sndtrack:  Allow longer freeform sounds
Let freeform waves be up to 64K long, instead of just 32K.  The "Mr. Ed"
application sends a 39,100-byte freeform wave to the Sound Driver, which
exceeds the maximum documented in Inside Macintosh, Volume I, as well as
our previous actual limit of 32,767 bytes.  The new cap is 65,535 bytes.

+++ sound/sndtrack/queues.hh
@@ -10,7 +10,7 @@
 struct queue_node
 {
 	queue_node*  next;
-	short        size;  // Any non-zero value indicates validity
+	int          size;  // Any non-zero value indicates validity
 	
 	bool valid() const  { return size != 0; }
